Implement BasicApp BasicAppConfig and BasicAppContext with their factories. One problem with one unit test !

+++ osgi-portal-api/src/main/java/fr/mby/portal/api/app/IAppContext.java
@@ -35,7 +35,7 @@ public interface IAppContext {
 
 	void log(String message, Throwable throwable, Object[]... objects);
 
-	int getBundleId();
+	long getBundleId();
 
 	String getWebContextPath();
